fix: need to investigate how to use effector state in tests
